+++ packages/types-codec/src/base/Option.ts
@@ -93,6 +93,11 @@ export class Option<T extends Codec> implements IOption<T> {
     }
   }
 
+  /** @deprecated Use $encodedLength */
+  public get encodedLength (): number {
+    return this.$encodedLength;
+  }
+
   /** @deprecated Use $initialU8aLength */
   public get initialU8aLength (): number | undefined {
     return this.$initialU8aLength;
@@ -117,9 +122,9 @@ export class Option<T extends Codec> implements IOption<T> {
   /**
    * @description The length of the value when encoded as a Uint8Array
    */
-  public get encodedLength (): number {
+  public get $encodedLength (): number {
     // boolean byte (has value, doesn't have) along with wrapped length
-    return 1 + this.#raw.encodedLength;
+    return 1 + this.#raw.$encodedLength;
   }
 
   /**
@@ -247,7 +252,7 @@ export class Option<T extends Codec> implements IOption<T> {
       return this.#raw.toU8a(true);
     }
 
-    const u8a = new Uint8Array(this.encodedLength);
+    const u8a = new Uint8Array(this.$encodedLength);
 
     if (this.isSome) {
       u8a.set([1]);