+++ packages/types-codec/src/extended/Map.ts
@@ -118,6 +118,11 @@ export class CodecMap<K extends Codec = Codec, V extends Codec = Codec> extends 
     this.#type = type;
   }
 
+  /** @deprecated Use $encodedLength */
+  public get encodedLength (): number {
+    return this.$encodedLength;
+  }
+
   /** @deprecated Use $initialU8aLength */
   public get initialU8aLength (): number | undefined {
     return this.$initialU8aLength;
@@ -126,11 +131,11 @@ export class CodecMap<K extends Codec = Codec, V extends Codec = Codec> extends 
   /**
    * @description The length of the value when encoded as a Uint8Array
    */
-  public get encodedLength (): number {
+  public get $encodedLength (): number {
     let len = compactToU8a(this.size).length;
 
     for (const [k, v] of this.entries()) {
-      len += k.encodedLength + v.encodedLength;
+      len += k.$encodedLength + v.$encodedLength;
     }
 
     return len;