+++ packages/types/src/ethereum/LookupSource.ts
@@ -69,7 +69,7 @@ export class GenericEthereumLookupSource extends AbstractBase<GenericEthereumAcc
   /**
    * @description The length of the value when encoded as a Uint8Array
    */
-  public override get encodedLength (): number {
+  public override get $encodedLength (): number {
     const rawLength = this._rawLength;
 
     return rawLength + (
@@ -86,7 +86,7 @@ export class GenericEthereumLookupSource extends AbstractBase<GenericEthereumAcc
   protected get _rawLength (): number {
     return this.inner instanceof GenericAccountIndex
       ? GenericAccountIndex.calcLength(this.inner)
-      : this.inner.encodedLength;
+      : this.inner.$encodedLength;
   }
 
   /**